made changes according to request
