+++ src/model/Partnership.ts
@@ -1,0 +1,39 @@
+import { observable } from 'mobx';
+
+import { Editor, service } from './service';
+import { Organization } from './Organization';
+
+export enum PartnershipTypes {
+    sponsor = 'sponsor',
+    place = 'place',
+    media = 'media',
+    community = 'community',
+    device = 'device',
+    travel = 'travel',
+    vendor = 'vendor'
+}
+
+export interface Partnership {
+    id: string;
+    title: string;
+    // activity?:	Activity;
+    level: number;
+    organization: Organization;
+    type: string;
+    accounts: [];
+    verified: boolean;
+    default: false;
+    created_by?: Editor;
+    updated_by?: Editor;
+}
+
+export class PartnershipModel {
+    @observable
+    current: Partnership = {} as Partnership;
+
+    async getOne(id: string) {
+        const { body } = await service.get<Partnership>('partner-ships/' + id);
+
+        return (this.current = body);
+    }
+}
