+++ src/model/index.ts
@@ -1,4 +1,8 @@
 import { History } from 'cell-router/source';
+import { PartnershipModel } from './Partnership';
 
 export const history = new History();
+export const partnership = new PartnershipModel();
 export * from './service';
+export * from './Organization';
+export * from './Partnership';
